Bugfix (maybe): properly clean up old decodes
This requires some more thoughts -- the shape of the problem is that the
encoder and decoder are running out of band with each other (very
intentionally -- it's the whole point). We want to "keep up" with what
the encoder throws at us, but we also need to clean up our list of
in-progress (and finished!) decodes at some point -- we don't want it to
be unbounded.

The bug this fixes should manifest as files refusing to
progress/download -- the decoder (app) has already reserved the "stream
slot", finished a download, but failed to clean up the mess to let it be
reclaimed. The workaround for CFC is to close and re-open the app, which
will throw out all in progress downloads and start from a clean slate.

+++ src/lib/fountain/FountainMetadata.h
@@ -33,7 +33,7 @@ public:
 		to_uint8_arr(encode_id, size, d);
 	}
 
-	unsigned id() const
+	uint64_t id() const
 	{
 		return _data;
 	}
